Migrate apply code to TS
