+++ package.json
@@ -25,6 +25,7 @@
   },
   "homepage": "https://github.com/hackcambridge/hack-cambridge-website#readme",
   "dependencies": {
+    "@types/express": "^4.0.39",
     "@types/node": "^8.0.55",
     "adjective-adjective-animal": "^1.3.3",
     "autoprefixer-loader": "^3.1.0",
